Add a new function to the 2D incompressible code
Now we have a new input parameter "if_z_radial" under namelist "AEB". If it is true, we are assuming that the simulation domain is perpendicular to the radial direction so both x and y dimensions are expanding. If it is false (as in previous versions), x is the radial direction and y is the transverse direction.