2D incom: Improve restart module
Make it compatible with the out999.dat, which is the backup file output every certain wall-time. Now the user should rename out999.dat or another output file to restart.dat. Then the program will automatically determine the next output time that is compatible with the dtout. For instance, if out999.dat was written at t=1.1 and dtout=0.5. the program will determine that next output time is t=1.5 instead of simply t=1.6.

Other changes include that the restart will not delete the old rms.dat file but will append to it.